improve security, view only assigned projects while adding activity, reformat

+++ app/routes.py
@@ -10,15 +10,18 @@ import datetime
 # Login/Register/Logout --------------------------------------------------------
 # ------------------------------------------------------------------------------
 
-# User check
-def check_user():
-    user = -1
-    if current_user.is_authenticated:
-        user = current_user.get_id()
-        return User.query.get(user)
-    else:
-        # TODO: User not logged in
-        return None
+# Get user
+
+def get_user():
+    return User.query.get(current_user.get_id())
+
+
+# Resource ownership check
+
+def check_ownership(user, required):
+    if required != user:
+        abort(403)
+
 
 # Registration
 @app.route("/register", methods=['GET', 'POST'])
@@ -33,6 +36,7 @@ def register():
         flash('Użytkownik utworzony', 'success')
         return redirect(url_for('login'))
     return render_template('register.html', title='Register', form=form)
+
 
 # Login
 @app.route('/', methods=['GET', 'POST'])
@@ -52,15 +56,12 @@ def login():
             error_message = 'Incorrect username or password.'
     return render_template('login.html', title='Login', form=form, error_message=error_message)
 
+
 # Logout
 @app.route('/logout')
 def logout():
     logout_user()
     return redirect(url_for('login'))
-
-
-
-
 
 
 # ------------------------------------------------------------------------------
@@ -69,29 +70,31 @@ def logout():
 
 # Activities list
 @app.route('/activities')
-#@login_required    # TODO: Uncomment in final
+@login_required
 def activities():
     current_view = 'activities'
-    u = check_user()
+    u = get_user()
 
-    if (u.role == 'pracownik'):
+    if u.role == 'pracownik':
         activityList = Activity.query.filter(Activity.user_id == u.id)
-    elif (u.role == 'klient'):
+    elif u.role == 'klient':
         activityList = db.session.query(Activity, Project).filter(Project.supervisor == u.id).all()
-    elif (u.role == 'kierownik'):
+    elif u.role == 'kierownik':
         activityList = Activity.query.all()
-    return render_template('common/activity/activities-list.html', activities=activityList, u=u, current_view=current_view)
+    return render_template('common/activity/activities-list.html', activities=activityList, u=u,
+                           current_view=current_view)
+
 
 # Activity add
 @app.route('/activities/add', methods=['GET', 'POST'])
 @app.route('/activities/add/<task_id>/', methods=['GET', 'POST'])
-#@login_required    # TODO: Uncomment in final
-def activity_add(task_id = None):
+@login_required
+def activity_add(task_id=None):
     form = AddActivityForm(task=task_id)
     current_view = 'activities-add'
-    u = check_user()
+    u = get_user()
 
-    if (u.role == 'pracownik'):
+    if u.role == 'pracownik':
         if form.validate_on_submit():
             user = current_user.get_id()
             activity = Activity(date=form.date.data, description=form.description.data, user_id=user,
@@ -101,24 +104,40 @@ def activity_add(task_id = None):
             db.session.commit()
             flash('Aktywność utworzona', 'success')
             return redirect(url_for('activities'))
-        return render_template('common/activity/activity-add.html', u=u, current_view=current_view, form=form, task_id = task_id)
+        return render_template('common/activity/activity-add.html', u=u, current_view=current_view, form=form,
+                               task_id=task_id)
     else:
         # TODO: Co w przypadku klienta i kierownika wchodzących na formularz dodania czynnosci?
         abort(403)
 
+
 # Delete activity
 @app.route('/projects/<int:project_id>/<int:task_id>/<int:activity_id>/delete', methods=['GET', 'POST'])
-#@login_required    # TODO: Uncomment in final
-def activity_delete(project_id = None, task_id = None, activity_id = None):
+@login_required
+def activity_delete(project_id=None, task_id=None, activity_id=None):
+    u = get_user()
+    if u.role == 'pracownik':
+        activity = Activity.query.get_or_404(activity_id)
+        check_ownership(int(current_user.get_id()), activity.user_id)
+    elif u.role == 'klient':
+        project = Project.query.get_or_404(project_id)
+        check_ownership(u.id, project.client)
+    elif u.role == 'kierownik':
+        project = Project.query.get_or_404(project_id)
+        check_ownership(u.id, project.supervisor)
+
     Activity.query.filter(Activity.id == activity_id).delete()
     db.session.commit()
     return redirect(url_for('projects'))
 
+
 # Accept activity - supervisor
 @app.route('/projects/<int:project_id>/<int:task_id>/<int:activity_id>/accept_supervisor/<int:state>')
-#@login_required    # TODO: Uncomment in final
-def activity_accept_supervisor(project_id = None, task_id = None, activity_id = None, state=None):
-    u = check_user()
+@login_required
+def activity_accept_supervisor(project_id=None, task_id=None, activity_id=None, state=None):
+    u = get_user()
+    project = Project.query.get_or_404(project_id)
+    check_ownership(u.id, project.supervisor)
     if u.role == 'kierownik':
         activity = Activity.query.filter(Activity.id == activity_id).first()
         if state == 0:
@@ -128,11 +147,14 @@ def activity_accept_supervisor(project_id = None, task_id = None, activity_id = 
         db.session.commit()
     return redirect(url_for('projects'))
 
+
 # Accept activity - customer
 @app.route('/projects/<int:project_id>/<int:task_id>/<int:activity_id>/accept_client/<int:state>')
-#@login_required    # TODO: Uncomment in final
-def activity_accept_client(project_id = None, task_id = None, activity_id = None, state=None):
-    u = check_user()
+@login_required
+def activity_accept_client(project_id=None, task_id=None, activity_id=None, state=None):
+    u = get_user()
+    project = Project.query.get_or_404(project_id)
+    check_ownership(u.id, project.client)
     if u.role == 'klient':
         activity = Activity.query.filter(Activity.id == activity_id).first()
         if state == 0:
@@ -143,24 +165,23 @@ def activity_accept_client(project_id = None, task_id = None, activity_id = None
     return redirect(url_for('projects'))
 
 
-
-
-
 # ------------------------------------------------------------------------------
 # Tasks ------------------------------------------------------------------------
 # ------------------------------------------------------------------------------
 
 # Add new task
 @app.route('/projects/<int:project_id>/task/add', methods=['GET', 'POST'])
-#@login_required    # TODO: Uncomment in final
+@login_required
 def task_add(project_id):
     form = TaskForm()
     current_view = 'task-add'
-    u = check_user()
+    u = get_user()
+    project = Project.query.get_or_404(project_id)
+    check_ownership(u.id, project.supervisor)
     project = Project.query.get_or_404(project_id)
     form.project_id = project.id
 
-    if (u.role == 'kierownik'):
+    if u.role == 'kierownik':
         if form.validate_on_submit():
             task = Task(name=form.name.data, description=form.description.data, project=project_id)
             db.session.add(task)
@@ -172,17 +193,16 @@ def task_add(project_id):
     else:
         abort(403)
 
+
 # Delete task
 @app.route('/projects/<int:project_id>/<int:task_id>/delete', methods=['GET', 'POST'])
-#@login_required    # TODO: Uncomment in final
-def task_delete(project_id = None, task_id = None):
+@login_required
+def task_delete(project_id=None, task_id=None):
+    project = Project.query.get_or_404(project_id)
+    check_ownership(int(current_user.get_id()), project.supervisor)
     Task.query.filter(Task.id == task_id).delete()
     db.session.commit()
     return redirect(url_for('projects'))
-
-
-
-
 
 
 # ------------------------------------------------------------------------------
@@ -191,11 +211,11 @@ def task_delete(project_id = None, task_id = None):
 
 # Single project view
 @app.route('/projects/<int:project_id>', methods=['GET', 'POST'])
-#@login_required    # TODO: Uncomment in final
+@login_required
 def projects_view(project_id):
     project = Project.query.get_or_404(project_id)
     current_view = 'projects-view'
-    u = check_user()
+    u = get_user()
 
     # Employee assign form
     form = EmployeeAssignForm()
@@ -208,38 +228,41 @@ def projects_view(project_id):
             # TODO: Clear form and assign properly
             return redirect(url_for('projects'))
 
-    return render_template('common/project/project-view.html', u=u, project=project, current_view=current_view, form=form)
+    return render_template('common/project/project-view.html', u=u, project=project, current_view=current_view,
+                           form=form)
+
 
 # Projects list
 @app.route('/projects')
-#@login_required    # TODO: Uncomment in final
+@login_required
 def projects():
     current_view = 'projects'
-    u = check_user()
+    u = get_user()
 
-    if(u.role == 'kierownik'):
+    if u.role == 'kierownik':
         projectlist = Project.query.filter(Project.supervisor == u.id)
 
-    if(u.role == 'klient'):
+    if u.role == 'klient':
         projectlist = Project.query.filter(Project.client == u.id)
 
     # Zmienic komentarz na czas developmentu jesli chcemy wyswietlac wszystkie projekty, a nie
     #   tylko te przypisane dla pracownika
-    if(u.role == 'pracownik'):
-        #projectlist = Project.query.all()
+    if u.role == 'pracownik':
+        # projectlist = Project.query.all()
         projectlist = [p.assigned_project for p in u.assigned_projects]
 
     return render_template('common/project/project-list.html', projects=projectlist, u=u, current_view=current_view)
 
+
 # Project add
 @app.route('/projects/add', methods=['GET', 'POST'])
-#@login_required    # TODO: Uncomment in final
+@login_required
 def project_add():
     form = ProjectForm()
     current_view = 'projects-add'
-    u = check_user()
+    u = get_user()
 
-    if (u.role == 'kierownik'):
+    if u.role == 'kierownik':
         if form.validate_on_submit():
             c = form.client.data
             if c == "":
@@ -255,17 +278,16 @@ def project_add():
     else:
         abort(403)
 
+
 # Delete project
 @app.route('/projects/<int:project_id>/delete', methods=['GET', 'POST'])
-#@login_required    # TODO: Uncomment in final
-def project_delete(project_id = None):
+@login_required
+def project_delete(project_id=None):
+    project = Project.query.get_or_404(project_id)
+    check_ownership(int(current_user.get_id()), project.supervisor)
     Project.query.filter(Project.id == project_id).delete()
     db.session.commit()
     return redirect(url_for('projects'))
-
-
-
-
 
 
 # ------------------------------------------------------------------------------
@@ -274,23 +296,24 @@ def project_delete(project_id = None):
 
 # Employee list
 @app.route('/employees')
-#@login_required    # TODO: Uncomment in final
+@login_required
 def employees():
     current_view = 'employees'
-    u = check_user()
+    u = get_user()
 
-    if (u.role == 'kierownik'):
+    if u.role == 'kierownik':
         employeeList = User.query.all()
         return render_template('admin/admin-employee-list.html', employees=employeeList, u=u, current_view=current_view)
     else:
         # TODO unauthantized
         return 0
 
+
 # Invoice printing
 @app.route('/projects/<int:project_id>/invoice', methods=['GET', 'POST'])
-#@login_required    # TODO: Uncomment in final
+@login_required
 def invoice_view(project_id):
     project = Project.query.get_or_404(project_id)
     current_view = 'invoice-view'
-    u = check_user()
+    u = get_user()
     return render_template('common/project/invoice-view.html', u=u, project=project, current_view=current_view)
